+++ tests/rules/test_PartialWildcardPrincipal.py
@@ -1,5 +1,6 @@
 from pytest import fixture
 
+from cfripper.config.config import Config
 from cfripper.rules import PartialWildcardPrincipalRule
 from tests.utils import get_cfmodel_from
 
@@ -12,6 +13,18 @@ def good_template():
 @fixture()
 def bad_template():
     return get_cfmodel_from("rules/PartialWildcardPrincipalRule/bad_template.json").resolve()
+
+
+@fixture()
+def intra_account_root_access():
+    return get_cfmodel_from("rules/PartialWildcardPrincipalRule/intra_account_root_access.yml").resolve()
+
+
+@fixture()
+def aws_elb_allow_template():
+    return get_cfmodel_from("rules/PartialWildcardPrincipalRule/aws_elb_template.yml").resolve(
+        extra_params={"AWS::Region": "ap-southeast-1"}
+    )
 
 
 def test_no_failures_are_raised(good_template):
@@ -28,19 +41,36 @@ def test_failures_are_raised(bad_template):
     result = rule.invoke(bad_template)
 
     assert not result.valid
-    assert len(result.failed_rules) == 4
+    assert len(result.failed_rules) == 2
     assert len(result.failed_monitored_rules) == 0
     assert result.failed_rules[0].rule == "PartialWildcardPrincipalRule"
-    assert result.failed_rules[0].reason == "PolicyA contains an unknown principal: 123445"
+    assert (
+        result.failed_rules[0].reason == "PolicyA should not allow wildcard in principals or account-wide principals "
+        "(principal: 'arn:aws:iam::123445:12345*')"
+    )
     assert result.failed_rules[1].rule == "PartialWildcardPrincipalRule"
     assert (
         result.failed_rules[1].reason == "PolicyA should not allow wildcard in principals or account-wide principals "
-        "(principal: 'arn:aws:iam::123445:12345*')"
-    )
-    assert result.failed_rules[2].rule == "PartialWildcardPrincipalRule"
-    assert result.failed_rules[2].reason == "PolicyA contains an unknown principal: 123445"
-    assert result.failed_rules[3].rule == "PartialWildcardPrincipalRule"
-    assert (
-        result.failed_rules[3].reason == "PolicyA should not allow wildcard in principals or account-wide principals "
         "(principal: 'arn:aws:iam::123445:root')"
     )
+
+
+def test_failures_for_correct_account_ids(intra_account_root_access):
+    rule = PartialWildcardPrincipalRule(Config(aws_account_id="123456789012"))
+    result = rule.invoke(intra_account_root_access)
+
+    assert not result.valid
+    assert len(result.failed_rules) == 1
+    assert len(result.failed_monitored_rules) == 0
+    assert result.failed_rules[0].rule == "PartialWildcardPrincipalRule"
+    assert (
+        result.failed_rules[0].reason
+        == "AccLoadBalancerAccessLogBucketPolicy should not allow wildcard in principals or account-wide principals "
+        "(principal: 'arn:aws:iam::123456789012:root')"
+    )
+
+
+def test_aws_elb_allow_template(aws_elb_allow_template):
+    rule = PartialWildcardPrincipalRule(None)
+    result = rule.invoke(aws_elb_allow_template)
+    assert result.valid