+++ CHANGELOG.md
@@ -5,10 +5,16 @@ All notable changes to this project will be documented in this file.
 ### Improvements
 - Implements `pluggy` https://github.com/pytest-dev/pluggy to enable dynamic rule loading.
 
-## [0.23.2] - 2021-02-01
+## [0.23.2] - 2021-02-04
+### Bugfix
+- `GenericWildcardPrincipalRule` to ignore account IDs where full or partial wildcard is required in the Principal.
+These accounts should be AWS Service Accounts defined in the config.
+- Fix CLI flag `--rules-config-file`
+
 ### Improvements
 - Update `ResourceSpecificRule` to allow for certain resources to be excluded. In particular, the
 `PrivilegeEscalationRule` will now no longer be invoked for `S3BucketPolicy` resources.
+- Add rules config for Kinesis Data Firehose IPs that can be applied
 
 ## [0.23.1] - 2021-01-26
 ### Improvements
@@ -34,7 +40,6 @@ All notable changes to this project will be documented in this file.
 ## [0.21.1] - 2020-12-9
 ### Improvements
 - Add SNS actions that only allow wildcards
-
 
 ## [0.21.0] - 2020-11-30
 ### Improvements